Ensure calculations return float values
Converted various calculations to float type for consistency.

+++ app.py
@@ -35,24 +35,24 @@ if tickers_input:
         log_returns = np.log(prices/prices.shift(1)).dropna()
         returns_dict[ticker] = log_returns
 
-        mu = log_returns.mean()*252
-        sigma = log_returns.std()*np.sqrt(252)
-
-        S0 = prices.iloc[-1]
+        mu = float(log_returns.mean()*252)
+        sigma = float(log_returns.std()*np.sqrt(252))
+        
+        S0 = float(prices.iloc[-1])
         T = forecast_days/252
-
-        expected_price = S0*np.exp(mu*T)
-
+        
+        expected_price = float(S0*np.exp(mu*T))
+        
         tp_price = S0*(1+tp_percent)
         sl_price = S0*(1-sl_percent)
-
+        
         d_tp = (np.log(tp_price/S0)-(mu-0.5*sigma**2)*T)/(sigma*np.sqrt(T))
-        prob_tp = 1-norm.cdf(d_tp)
-
+        prob_tp = float(1-norm.cdf(d_tp))
+        
         d_sl = (np.log(sl_price/S0)-(mu-0.5*sigma**2)*T)/(sigma*np.sqrt(T))
-        prob_sl = norm.cdf(d_sl)
-
-        emotional_index = (expected_price/S0-1)*prob_tp - sigma*prob_sl
+        prob_sl = float(norm.cdf(d_sl))
+        
+        emotional_index = float((expected_price/S0-1)*prob_tp - sigma*prob_sl)
 
         if emotional_index > 0.05:
             emotion = "🟢 Otimista"
